refactor start endless game redux logic into action
