+++ src/features/dialog-manager/dialogs/endless-game-start/index.js
@@ -1,125 +1,31 @@
 import React, { useEffect } from 'react';
+import { connect }          from 'react-redux';
 
 import Button           from '../../../../components/button';
 import Dialog           from '../../../../components/dialog';
-import items            from '../../../../data/items';
-import generateMap      from '../../../map/random-map-gen/generate-map';
-import generateMonsters from '../../../map/random-map-gen/generate-monsters';
-import exploreTiles     from '../../../../features/player/actions/explore-tiles';
-import uuidv4           from '../../../../modules/uuid-v4';
-
-import store from '../../../../config/store';
+import backToSelect     from '../../actions/back-to-select';
+import startEndlessGame from '../../../world/actions/start-endless-game';
 
 import './styles.scss';
 
-const EndlessGameStart = () => {
+const EndlessGameStart = ({ backToSelect, startEndlessGame }) => {
 
   useEffect(() => {
     window.addEventListener('keydown', handleKeyPress);
     return () => {
       window.removeEventListener('keydown', handleKeyPress);
     }
-  }, []);  // we pass empty array as the second param to make this only call on mount and not on any updates
+  }, []);
 
   function handleKeyPress(event) {
     // case for 'enter' and 'space' key
     if(event.keyCode === 13 || event.keyCode === 32) {
-      handleGameStart();
+      startEndlessGame();
     }
   }
 
-  function handleGameStart() {
-    const { player, stats } = store.getState();
-    const floorNum = 1;
-    // generate a random map and id
-    const randomMap = generateMap(player.position, floorNum);
-    const mapId = uuidv4();
-
-    handleShowStartMessage();
-    handleLoadMap(mapId, randomMap, floorNum);
-    handleLoadPlayerSight(player);
-    handleLoadMonsters(floorNum, mapId, randomMap, player, stats);
-    handleLoadStartingItems();
-  }
-
-  function handleShowStartMessage() {
-    store.dispatch({
-      type: 'PAUSE',
-      payload: {
-        pause: true,
-        gameText: {
-          title: `As you stare into the dark dungeon, it greets you with a cold chill... and a message...`,
-          body: `"JOURNEY ONE HUNDRED FLOORS AND ALL WILL BE GRANTED"`
-        }
-      }
-    });
-  }
-
-  function handleLoadMap(mapId, randomMap, floorNum) {
-    // set map tiles for start map
-    store.dispatch({
-      type: 'SET_START_MAP',
-      payload: {
-        startMap: mapId,
-        gameMode: 'endless',
-        floorNum: floorNum
-      }
-    });
-
-    store.dispatch({
-      type: 'ADD_TILES',
-      payload: { tiles: randomMap }
-    });
-
-    store.dispatch({
-      type: 'ADD_RANDOM_MAP',
-      payload: {
-        tiles: randomMap,
-        id: mapId
-      }
-    });
-  }
-
-  function handleLoadPlayerSight(player) {
-    exploreTiles(player.position);
-  }
-
-  function handleLoadMonsters(floorNum, mapId, randomMap, player, stats) {
-    // load initial (random) monsters
-    store.dispatch({
-      type: 'ADD_MONSTERS',
-      payload: {
-        monsters: generateMonsters(floorNum, randomMap, player.position, stats.level),
-        map: mapId
-      }
-    });
-  }
-
-  function handleLoadStartingItems() {
-    // give the player a rusty sword
-    store.dispatch({
-      type: 'GET_ITEM',
-      payload: items.weapons.RustySword
-    });
-    // and equip it
-    store.dispatch({
-      type: 'EQUIP_ITEM',
-      payload: store.getState().inventory.items[0]
-    });
-  }
-
-  function goBack() {
-    store.dispatch({
-      type: 'PAUSE',
-      payload: {
-        pause: true,
-        gameStart: true
-      }
-    });
-  }
-
   return(
-    <Dialog goBack={goBack}>
+    <Dialog goBack={backToSelect}>
 
       <span className='flex-row endless-start__title'>
         {'Endless Mode'}
@@ -131,7 +37,7 @@ const EndlessGameStart = () => {
 
       <div className='flex-column endless-start__button'>
         <Button
-          onClick={handleGameStart}
+          onClick={startEndlessGame}
           icon='compass'
           title={'Explore Dungeon'} />
       </div>
@@ -139,4 +45,6 @@ const EndlessGameStart = () => {
   );
 }
 
-export default EndlessGameStart;
+const actions = { backToSelect, startEndlessGame };
+
+export default connect(null, actions)(EndlessGameStart);
